Bug 433409 - ScriptEngine can't be retrieved when using non-Hotsport/OpenJDK Java
The method ScriptEngineManager(ClassLoader) only searches in the
provided class loader. Which is the behavior we do not want. Instead
we want the behavior from ScriptEngineManager() which has all script
engine that are present in the platform.

However the Hotspot Java (up to 7) also adds the script engines present
in the platform to the list of found script engines.

Under OSGi no script engines will be found using the
ScriptEngineManager(ClassLoader) constructor since the service locator
feature won't work.