allow disabling the write dialog of detail views
the detail view had their own write confirmation dialog. since we
now have server triggered write dialogs, this would cause duplicate
confirmation dialogs for one action.

it is now possible to specify if a detail view should enable these
client side confirmation dialogs or not. The default is to use
these dialogs since this is the behavior up to now.