Fix a possible NPE and a possible NOOP
This change fixes a NPE when the change model is stopped but was never
started.

In addition is fixes a possible no-op situation when the model gets
disposed and the write model internally set to null, then the write
model reference of the change model is null, and thus the scheduled
actions cannot be executed anymore. This is fixed by making a copy
of the write model when the job is scheduled.