make ConnectionInformation immutable; closes #3
also adds validation of port (which was the underlying cause of this
bug). We need a valid port anyway, so this shouldn't be a problem.

unfortunately this is an backwards incompatible change, but essentially
everywhere fromUri is used anyway, and the singular place where the
mutable user/password is used, this is fixed in this commit as well

Signed-off-by: Juergen Rose <juergen.rose@ibh-systems.com>